Move SqlServerDacpacDatabaseModelFactory to its own package
