+++ src/GUI/efreveng/ReverseEngineer/ServiceProviderBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using EntityFrameworkCore.Scaffolding.Handlebars;
+using ErikEJ.EntityFrameworkCore.SqlServer.Scaffolding;
 using Microsoft.EntityFrameworkCore.Design;
 using Microsoft.EntityFrameworkCore.Design.Internal;
 using Microsoft.EntityFrameworkCore.Scaffolding;
